# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartReserve — Smart Demand-Aware Offer Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List offers (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: businessId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create an offer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get offer details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offer with slots and business info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOffer
      tags: [offers]
      summary: Update an offer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteOffer
      tags: [offers]
      summary: Delete an offer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /offers/{id}/status:
    patch:
      operationId: updateOfferStatus
      tags: [offers]
      summary: Update offer status (pause/activate/cancel)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"

  # SLOTS
...
  /public/offers:
    get:
      operationId: listPublicOffers
      tags: [public]
      summary: List active offers for public portal
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Public offers with demand info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicOffer"
...
  /public/offers/{id}:
    get:
      operationId: getPublicOffer
      tags: [public]
      summary: Get public offer detail with slots and demand info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public offer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicOfferDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # AUTH
...
    Business:
      type: object
      required: [id, name, businessType, ownerName, phone, email, address, city, openingTime, closingTime]
      properties:
        id:
          type: integer
        name:
          type: string
        businessType:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        openingTime:
          type: string
        closingTime:
          type: string
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Offer:
      type: object
      required: [id, businessId, title, description, category, originalPrice, offerPrice, discountPercent, startDate, endDate, startTime, endTime, status]
      properties:
        id:
          type: integer
        businessId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        originalPrice:
          type: number
        offerPrice:
          type: number
        discountPercent:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        terms:
          type: ["string", "null"]
        status:
          type: string
        bannerUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    OfferInput:
      type: object
      required: [businessId, title, description, category, originalPrice, offerPrice, startDate, endDate, startTime, endTime]
      properties:
        businessId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        originalPrice:
          type: number
        offerPrice:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        terms:
          type: string
        status:
          type: string
        bannerUrl:
          type: string
    OfferUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        originalPrice:
          type: number
        offerPrice:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        terms:
          type: string
        status:
          type: string
        bannerUrl:
          type: string
    OfferStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
    OfferDetail:
      type: object
      required: [id, businessId, title, description, category, originalPrice, offerPrice, discountPercent, startDate, endDate, startTime, endTime, status, business, slots]
      properties:
        id:
          type: integer
        businessId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        originalPrice:
          type: number
        offerPrice:
          type: number
        discountPercent:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        terms:
          type: ["string", "null"]
        status:
          type: string
        bannerUrl:
          type: ["string", "null"]
        business:
          $ref: "#/components/schemas/Business"
        slots:
          type: array
          items:
            $ref: "#/components/schemas/Slot"
        createdAt:
          type: string
          format: date-time

    # SLOT
    Slot:
      type: object
      required: [id, offerId, slotDate, startTime, endTime, capacity, bookedCount, availableCount, status]
      properties:
        id:
          type: integer
        offerId:
          type: integer
        slotDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        capacity:
          type: integer
        bookedCount:
          type: integer
        availableCount:
          type: integer
        status:
          type: string
        occupancyPercent:
          type: number
        demandLevel:
          type: string
...
    PublicOffer:
      type: object
      required: [id, title, description, category, originalPrice, offerPrice, discountPercent, status, business, totalCapacity, totalBooked, occupancyPercent, demandLevel, availableSlots]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        originalPrice:
          type: number
        offerPrice:
          type: number
        discountPercent:
          type: number
        status:
          type: string
        bannerUrl:
          type: ["string", "null"]
        business:
          $ref: "#/components/schemas/Business"
        totalCapacity:
          type: integer
        totalBooked:
          type: integer
        occupancyPercent:
          type: number
        demandLevel:
          type: string
        availableSlots:
          type: integer
        recentBookings:
          type: integer
        isTrending:
          type: boolean
        isFillingFast:
          type: boolean
        recommendedSlotTime:
          type: ["string", "null"]
        endDate:
          type: string
        endTime:
          type: string
    PublicOfferDetail:
      type: object
      required: [id, title, description, category, originalPrice, offerPrice, discountPercent, status, business, slots, demandInfo]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        originalPrice:
          type: number
        offerPrice:
          type: number
        discountPercent:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        terms:
          type: ["string", "null"]
        status:
          type: string
        bannerUrl:
          type: ["string", "null"]
        business:
          $ref: "#/components/schemas/Business"
        slots:
          type: array
          items:
            $ref: "#/components/schemas/Slot"
        demandInfo:
          $ref: "#/components/schemas/DemandInfo"
    DemandInfo:
      type: object
      required: [recentBookings, occupancyPercent, demandLevel, isFillingFast, isTrending, recommendedSlotId, recommendedSlotTime]
      properties:
        recentBookings:
          type: integer
        occupancyPercent:
          type: number
        demandLevel:
          type: string
        isFillingFast:
          type: boolean
        isTrending:
          type: boolean
        recommendedSlotId:
          type: ["integer", "null"]
        recommendedSlotTime:
          type: ["string", "null"]
...