# lib/api-zod/src/generated/types/publicOffer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartReserve — Smart Demand-Aware Offer Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Business } from './business';

export interface PublicOffer {
  id: number;
  title: string;
  description: string;
  category: string;
  originalPrice: number;
  offerPrice: number;
  discountPercent: number;
  status: string;
  /** @nullable */
  bannerUrl?: string | null;
  business: Business;
  totalCapacity: number;
  totalBooked: number;
  occupancyPercent: number;
  demandLevel: string;
  availableSlots: number;
  recentBookings?: number;
  isTrending?: boolean;
  isFillingFast?: boolean;
  /** @nullable */
  recommendedSlotTime?: string | null;
  endDate?: string;
  endTime?: string;
}